updated code

+++ models/official/projects/maskformer/losses/inference.py
@@ -72,7 +72,7 @@ class PanopticInference:
                     
                     # isthing = self.is_thing_dict[self.cat_id_map[int(pred_class)]]
 
-                    binary_mask = tf.math.equal(cur_mask_ids, k)
+                    binary_mask = tf.math.equal(cur_mask_ids, tf.cast(k, tf.int64))
                     binary_mask = tf.cast(binary_mask, tf.int32)
                     
                     mask_area = tf.math.reduce_sum(binary_mask)
